refactor(api): deduplicate WeightedScoreConfig and clean up remaining import paths
refactor(api): deduplicate RAG index entities and consolidate import paths
