+++ api/core/rag/entities/__init__.py
@@ -1,12 +1,20 @@
 from core.rag.entities.citation_metadata import RetrievalSourceMetadata
 from core.rag.entities.context_entities import DocumentContext
+from core.rag.entities.event import DatasourceCompletedEvent, DatasourceErrorEvent, DatasourceProcessingEvent
+from core.rag.entities.index_entities import EconomySetting, EmbeddingSetting, IndexMethod
 from core.rag.entities.metadata_entities import Condition, MetadataFilteringCondition, SupportedComparisonOperator
 from core.rag.entities.processing_entities import ParentMode, PreProcessingRule, Rule, Segmentation
-from core.rag.entities.retrieval_settings import KeywordSetting, VectorSetting
+from core.rag.entities.retrieval_settings import KeywordSetting, VectorSetting, WeightedScoreConfig
 
 __all__ = [
     "Condition",
+    "DatasourceCompletedEvent",
+    "DatasourceErrorEvent",
+    "DatasourceProcessingEvent",
     "DocumentContext",
+    "EconomySetting",
+    "EmbeddingSetting",
+    "IndexMethod",
     "KeywordSetting",
     "MetadataFilteringCondition",
     "ParentMode",
@@ -16,4 +24,5 @@ __all__ = [
     "Segmentation",
     "SupportedComparisonOperator",
     "VectorSetting",
+    "WeightedScoreConfig",
 ]
