+++ api/core/workflow/nodes/knowledge_index/entities.py
@@ -1,10 +1,10 @@
-from typing import Literal, Union
+from typing import Union
 
 from graphon.entities.base_node_data import BaseNodeData
 from graphon.enums import NodeType
 from pydantic import BaseModel
 
-from core.rag.entities import KeywordSetting, VectorSetting
+from core.rag.entities import WeightedScoreConfig
 from core.rag.index_processor.index_processor_base import SummaryIndexSettingDict
 from core.rag.retrieval.retrieval_methods import RetrievalMethod
 from core.workflow.nodes.knowledge_index import KNOWLEDGE_INDEX_NODE_TYPE
@@ -17,32 +17,6 @@ class RerankingModelConfig(BaseModel):
 
     reranking_provider_name: str
     reranking_model_name: str
-
-
-class WeightedScoreConfig(BaseModel):
-    """
-    Weighted score Config.
-    """
-
-    vector_setting: VectorSetting
-    keyword_setting: KeywordSetting
-
-
-class EmbeddingSetting(BaseModel):
-    """
-    Embedding Setting.
-    """
-
-    embedding_provider_name: str
-    embedding_model_name: str
-
-
-class EconomySetting(BaseModel):
-    """
-    Economy Setting.
-    """
-
-    keyword_number: int
 
 
 class RetrievalSetting(BaseModel):
@@ -58,16 +32,6 @@ class RetrievalSetting(BaseModel):
     reranking_enable: bool = True
     reranking_model: RerankingModelConfig | None = None
     weights: WeightedScoreConfig | None = None
-
-
-class IndexMethod(BaseModel):
-    """
-    Knowledge Index Setting.
-    """
-
-    indexing_technique: Literal["high_quality", "economy"]
-    embedding_setting: EmbeddingSetting
-    economy_setting: EconomySetting
 
 
 class FileInfo(BaseModel):
