+++ api/core/workflow/nodes/knowledge_retrieval/entities.py
@@ -5,7 +5,7 @@ from graphon.enums import BuiltinNodeTypes, NodeType
 from graphon.nodes.llm.entities import ModelConfig, VisionConfig
 from pydantic import BaseModel, Field
 
-from core.rag.entities import Condition, KeywordSetting, MetadataFilteringCondition, VectorSetting
+from core.rag.entities import Condition, MetadataFilteringCondition, WeightedScoreConfig
 
 __all__ = ["Condition"]
 
@@ -17,15 +17,6 @@ class RerankingModelConfig(BaseModel):
 
     provider: str
     model: str
-
-
-class WeightedScoreConfig(BaseModel):
-    """
-    Weighted score Config.
-    """
-
-    vector_setting: VectorSetting
-    keyword_setting: KeywordSetting
 
 
 class MultipleRetrievalConfig(BaseModel):
