+++ api/services/entities/knowledge_entities/rag_pipeline_entities.py
@@ -46,23 +46,6 @@ class WeightedScoreConfig(BaseModel):
     keyword_setting: KeywordSetting | None
 
 
-class EmbeddingSetting(BaseModel):
-    """
-    Embedding Setting.
-    """
-
-    embedding_provider_name: str
-    embedding_model_name: str
-
-
-class EconomySetting(BaseModel):
-    """
-    Economy Setting.
-    """
-
-    keyword_number: int
-
-
 class RetrievalSetting(BaseModel):
     """
     Retrieval Setting.
@@ -76,16 +59,6 @@ class RetrievalSetting(BaseModel):
     reranking_enable: bool | None = True
     reranking_model: RerankingModelConfig | None = None
     weights: WeightedScoreConfig | None = None
-
-
-class IndexMethod(BaseModel):
-    """
-    Knowledge Index Setting.
-    """
-
-    indexing_technique: Literal["high_quality", "economy"]
-    embedding_setting: EmbeddingSetting
-    economy_setting: EconomySetting
 
 
 class KnowledgeConfiguration(BaseModel):
